// src/app/models/invoice/payment-search-view-model.ts
/**
 * Fee
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * Invoice Criteria ViewModel
 */
export interface PaymentSearchViewModel {
    sortBy: any;
    sortOrder: 0 | 1 | -1;
    pageNumber: number;
    pageSize: number;
    academicSessionId?: number;
    feeTermId?: number;
    boardId?: number;
    classId?: number;
    sectionId?: number;
    studentId?: any;
    invoiceNo?: string;
    receiptNo?: string;
    paymentStatusId?: number;
    admissionNo?: Array<string>;
    paymentDatesBegin?: Date;
    paymentDatesEnd?: Date;
    paymentType?: Array<string>;
    feeAccount?: Array<string>;
    paymentAmount?: Array<string>;
    invoiceAccounted?: Array<string>;
    balanceAmount?: Array<string>;

}